backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic-settings==2.5.2
email-validator==2.2.0
pymongo==4.8.0
motor==3.5.1
passlib==1.7.4
PyJWT==2.9.0
qdrant-client==1.11.1
sentence-transformers==3.1.1
scikit-learn==1.5.2
numpy==1.26.4
scipy==1.14.1
PyMuPDF==1.24.10
rapidocr-onnxruntime==1.3.24
Pillow==10.4.0
google-generativeai==0.8.3
httpx==0.27.2
tavily-python==0.5.0
python-dotenv==1.0.1
